# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KhataPro CRM API — digital ledger platform for small businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors for a business
      parameters:
        - name: business_id
          in: query
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
# ─── PURCHASES ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Vendor:
      type: object
      required: [id, business_id, name, created_at]
      properties:
        id:
          type: integer
        business_id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        gst_number:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    VendorInput:
      type: object
      required: [business_id, name]
      properties:
        business_id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gst_number:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gst_number:
          type: string
    VendorListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
# ── Purchase ──────────────────────────────────────────────────────────────  
...